requirements.txt
certifi
charset-normalizer
idna
mariadb
python-dotenv
requests
urllib3
